Refactor factory helpers
