+++ src/flatcollections-array/FlatArray/FlatArray.T/Factory/Factory.Explicit.From.FlatArray.cs
@@ -14,26 +14,11 @@ partial struct FlatArray<T>
 
     // TODO: Add the tests and make public
     internal static FlatArray<T> From(FlatArray<T> source, int start, int length)
-    {
-        InnerValidateRange();
-
-        return length == default
-            ? default
-            : new(InnerArrayHelper.CopySegment(source.items!, start, length), default);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        void InnerValidateRange()
-        {
-            if (InnerAllocHelper.IsSegmentWithinBounds(start, length, source.length))
-            {
-                return;
-            }
-            throw InnerExceptionFactory.SegmentOutsideBounds(start, length, source.length);
-        }
-    }
+        =>
+        InnerFactoryHelper.FromFlatArray(source, start, length);
 
     // TODO: Add the tests and make public
     internal static FlatArray<T> From(FlatArray<T>? source, int start, int length)
         =>
-        From(source.GetValueOrDefault(), start, length);
+        InnerFactoryHelper.FromFlatArray(source.GetValueOrDefault(), start, length);
 }
