+++ src/flatcollections-array/FlatArray/FlatArray.T/FlatArray.Slice.cs
@@ -5,5 +5,5 @@ partial struct FlatArray<T>
     // TODO: Add the tests and make public
     internal FlatArray<T> Slice(int start, int length)
         =>
-        From(this, start, length);
+        InnerFactoryHelper.FromFlatArray(this, start, length);
 }
